Support icons in Popover Items & actions.
