+++ src/components/overlays/Popover/Item/Item.tsx
@@ -2,12 +2,21 @@ import React from 'react'
 import { BodyText } from '../../../text'
 import { Box } from '../../../structure'
 import { Touchable } from '../../../base'
+import { IconName, IconProps, Icon } from '../../../images-and-icons'
 
 export interface ItemProps {
   /**
    * Label of the Item.
    */
   label: string
+  /**
+   * Optional Icon.
+   */
+  icon?: IconName
+  /**
+   * Color of the Icon.
+   */
+  iconColor?: IconProps['color']
   /**
    * Action to execute on click.
    */
@@ -17,9 +26,10 @@ export interface ItemProps {
 /**
  * Item shown in a Popover.
  */
-export const Item: React.FC<ItemProps> = ({ label, onSelect }) => (
+export const Item: React.FC<ItemProps> = ({ label, icon, iconColor, onSelect }) => (
   <Touchable onClick={onSelect}>
-    <Box paddingX="medium" paddingY="small">
+    <Box horizontal space="small" paddingX="medium" paddingY="small" align="center">
+      {icon ? <Icon name={icon} color={iconColor} /> : null}
       <BodyText>{label}</BodyText>
     </Box>
   </Touchable>
