+++ src/components/overlays/Popover/Popover.tsx
@@ -1,30 +1,18 @@
-import _ from 'lodash'
-import React, { useMemo, useState, useCallback, useRef } from 'react'
-import { View, TouchableWithoutFeedback, ViewProps, LayoutRectangle } from 'react-native'
-import { useSafeArea } from 'react-native-safe-area-context'
-import { TextAction } from '../../actions'
+import React, { useMemo, useRef } from 'react'
+import { View, TouchableWithoutFeedback, LayoutRectangle } from 'react-native'
+import { TextWithOptionalIconAction } from '../../actions'
 import { useStyles } from '../../../theme'
 import { shameStyles } from '../../../theme/shame-styles'
-import {
-  useAppProviderDimensions,
-  useAppProviderPosition,
-} from '../../dev/SizeProvider/AppProviderSizeProvider'
+import { useAppProviderPosition } from '../../dev/SizeProvider/AppProviderSizeProvider'
 import { Item, ItemProps } from './Item/Item'
-import {
-  PopoverPlacement,
-  isTop,
-  isBottom,
-  isStart,
-  isEnd,
-  isLeft,
-  isRight,
-} from './popover-placement'
+import { PopoverPlacement } from './popover-placement'
+import { PopoverView } from './PopoverView/PopoverView'
 
 export interface PopoverWithActions {
   /**
    * Actions to display in the Popover.
    */
-  actions: TextAction[]
+  actions: TextWithOptionalIconAction[]
   children?: never
 }
 
@@ -74,6 +62,9 @@ const { backdrop } = shameStyles.popover
 
 const LAYOUT_ZERO: LayoutRectangle = { x: 0, y: 0, width: 0, height: 0 }
 
+/**
+ * Popovers are small overlays that open on demand. They are meant to access additional content without cluttering the screen.
+ */
 export const Popover: React.FC<PopoverProps> & { Item: typeof Item } = ({
   open,
   anchor,
@@ -127,8 +118,11 @@ export const Popover: React.FC<PopoverProps> & { Item: typeof Item } = ({
 export type PopoverItemProps = ItemProps
 Popover.Item = Item
 
+/**
+ * Get the Popover content (either through `children` or `actions`).
+ */
 const useContent = (
-  actions: TextAction[] | undefined,
+  actions: TextWithOptionalIconAction[] | undefined,
   children: React.ReactNode | undefined,
 ): React.ReactNode =>
   useMemo(() => {
@@ -141,154 +135,10 @@ const useContent = (
     let content: React.ReactNode = children
 
     if (actions) {
-      content = actions.map(({ label, action }, i) => (
-        <Item key={i} label={label} onSelect={action} />
+      content = actions.map(({ label, icon, color, action }, i) => (
+        <Item key={i} label={label} icon={icon} iconColor={color} onSelect={action} />
       ))
     }
 
     return content
   }, [actions, children])
-
-interface PopoverViewProps {
-  children: React.ReactNode
-  placement: PopoverPlacement
-  matchWidth: boolean
-  anchorLayout: LayoutRectangle
-}
-
-const PopoverView: React.FC<PopoverViewProps> = ({
-  children,
-  placement,
-  matchWidth,
-  anchorLayout,
-}) => {
-  const styles = useStyles(theme => ({
-    container: {
-      position: 'absolute',
-      top: 0,
-      left: 0,
-      alignItems: 'flex-start',
-    },
-    popover: {
-      backgroundColor: theme.colors.fill.background.lighter,
-      ...theme.elevation.z8,
-    },
-    popoverNotYetLayout: {
-      opacity: 0,
-    },
-    popoverMatchWidth: {
-      left: 0,
-      right: 0,
-    },
-  }))
-  const insets = useSafeArea()
-  const { width: windowWidth, height: windowHeight } = useAppProviderDimensions()
-
-  const [layout, setLayout] = useState<LayoutRectangle | null>(null)
-  const onLayout = useCallback<Required<ViewProps>['onLayout']>(
-    ({ nativeEvent }) => setLayout(nativeEvent.layout),
-    [setLayout],
-  )
-
-  // Calculate the offset of the Popover relative to parent according to placement
-  const placementOffset = usePlacementOffset(placement, layout, anchorLayout)
-
-  // Corrects the placement to be inside window bounds
-  const correctedPlacementOffset = useOffsetsCorrectionToBeInWindow(
-    placementOffset,
-    layout,
-    anchorLayout,
-    windowWidth,
-    windowHeight,
-  )
-
-  return (
-    <View
-      style={[
-        styles.container,
-        {
-          width: windowWidth - insets.left - insets.right,
-          height: windowHeight - insets.top - insets.bottom,
-        },
-        {
-          marginTop: insets.top,
-          marginBottom: insets.bottom,
-          marginLeft: insets.left,
-          marginRight: insets.right,
-        },
-      ]}
-      pointerEvents="none"
-    >
-      <View
-        style={[
-          styles.popover,
-          layout === null
-            ? styles.popoverNotYetLayout
-            : {
-                transform: [
-                  { translateY: correctedPlacementOffset.y },
-                  { translateX: correctedPlacementOffset.x },
-                ],
-              },
-          { maxWidth: windowWidth, maxHeight: windowHeight },
-          matchWidth ? { width: anchorLayout.width } : null,
-        ]}
-        onLayout={onLayout}
-      >
-        {children}
-      </View>
-    </View>
-  )
-}
-
-const usePlacementOffset = (
-  placement: PopoverPlacement,
-  popoverLayout: LayoutRectangle | null,
-  anchorLayout: LayoutRectangle,
-): Offset =>
-  useMemo(
-    () => ({
-      x: isLeft(placement)
-        ? -(popoverLayout?.width ?? 0)
-        : isRight(placement)
-        ? anchorLayout.width
-        : isStart(placement)
-        ? 0
-        : isEnd(placement)
-        ? -(popoverLayout?.width ?? 0) + anchorLayout.width
-        : (anchorLayout.width - (popoverLayout?.width ?? 0)) / 2, // Top or bottom centered
-      y: isTop(placement)
-        ? -(popoverLayout?.height ?? 0)
-        : isBottom(placement)
-        ? anchorLayout?.height ?? 0
-        : isStart(placement)
-        ? 0
-        : isEnd(placement)
-        ? -(popoverLayout?.height ?? 0) + anchorLayout.height
-        : (anchorLayout.height - (popoverLayout?.height ?? 0)) / 2, // Left or right centered
-    }),
-    [placement, popoverLayout, anchorLayout],
-  )
-
-const useOffsetsCorrectionToBeInWindow = (
-  offset: Offset,
-  popoverLayout: LayoutRectangle | null,
-  anchorLayout: LayoutRectangle,
-  windowWidth: number,
-  windowHeight: number,
-): Offset =>
-  useMemo(() => {
-    if (popoverLayout === null) {
-      return { x: 0, y: 0 }
-    }
-
-    return {
-      x: _.clamp(offset.x, -anchorLayout.x, -anchorLayout.x + windowWidth - popoverLayout.width),
-      y: _.clamp(offset.y, -anchorLayout.y, -anchorLayout.y + windowHeight - popoverLayout.height),
-    }
-  }, [popoverLayout, anchorLayout, offset, windowWidth, windowHeight])
-
-interface Offset {
-  x: number
-  y: number
-}
